src/tests/test_data_processing.py
import unittest
from src import data_processing

class TestDataProcessing(unittest.TestCase):

    def setUp(self):
        self.dataSet = [
            # Add some sample data for testing
        ]
        self.processedData = None

    def test_processData(self):
        self.processedData = data_processing.processData(self.dataSet)
        self.assertIsNotNone(self.processedData, "Data processing failed.")

    def test_ScientificDataSchema(self):
        schema = data_processing.ScientificDataSchema()
        validation = schema.validate(self.processedData)
        self.assertTrue(validation, "Data does not match the schema.")

    def test_dataProcessed(self):
        self.assertTrue(data_processing.dataProcessed, "Data processed message not sent.")

if __name__ == '__main__':
    unittest.main()